extended life

+++ src/core/Odin.Core/IClientRegistration.cs
@@ -7,7 +7,7 @@ public interface IClientRegistration
     public Guid Id { get; }
     public string IssuedTo { get; }
     public int Type { get; }
-    public long TimeToLiveSeconds { get; }
+    public int TimeToLiveSeconds { get; }
 
     /// <summary>
     /// Categorizes this registration, giving you a way to get a list by something like AppId
