+++ src/services/Odin.Services/Membership/YouAuth/YouAuthDomainClient.cs
@@ -28,7 +28,7 @@ public sealed class YouAuthDomainClient : IClientRegistration
 
     public string IssuedTo => this.Domain.DomainName;
     public int Type => 408;
-    public long TimeToLiveSeconds { get; set; } 
+    public int TimeToLiveSeconds { get; set; } 
     public Guid CategoryId { get; } = Guid.Parse("83742ae7-e66d-45e6-82a6-6a003c960b39");
 
     public string GetValue()
